worker/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.6.0
pymupdf==1.24.0
pytesseract==0.3.10
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers>=2.2.0
